improve CellTypeRef types, access CellType more directly
Don't attempt to make CellType a nice high-level public API.

New plan: later a separate CellTypeConfig struct shall be used as
higher-level API, where cell types and references are strings. This can
be serde'd to/from yaml and json. This CellTypeConfig can then be
converted to the low-level CellTypes (but not the other way around).