hyperparam tuning with ASHA
Fun fact: having 300 independent runs and stopping bad ones early
("restarts" in CMA-ES terms, I think?) was the more important effect of
using ASHA, not nailing the hyperparams.

(In this toy problem, so far. Probably the fitness function is very
misleading. Also, the space of possible policies is currently so small
that a LUT (state -> action prob) would be a better fit than a NN.)